pyproject.toml
[project]
name = "2025-fastapi-seminar-template-3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiomysql>=0.3.2",
    "aiosqlite>=0.21.0",
    "alembic>=1.16.5",
    "argon2-cffi>=25.1.0",
    "authlib>=1.6.4",
    "fastapi[standard]>=0.117.1",
    "pydantic-settings>=2.10.1",
    "pymysql>=1.1.2",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.3.0",
    "sqlalchemy>=2.0.43",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
norecursedirs = ["legacy_tests"]
